requirements.txt
certifi~=2021.10.8
click~=8.1.2
dnspython~=2.2.1
Flask~=2.1.1
importlib-metadata~=4.11.3
itsdangerous~=2.1.2
Jinja2~=3.1.1
MarkupSafe~=2.1.1
pydantic~=1.9.0
pymongo~=4.1.1
python-dotenv~=0.20.0
typing_extensions~=4.2.0
Werkzeug~=2.1.1
zipp~=3.8.0
